Update examples to be v.0.3.0+ compliant (#69)
* Add .gitignore file

This commit adds a .gitignore file to ignore tracking changes in
the .ipynb_checkpoints directory.

* Update plot_confusion_matrix example notebook (#62)

This commit updates this example to be compliant on v.0.3.0. Using
the classifier_factory has been removed, and instead plots the
confusion matrix using scikitplot.metrics.plot_confusion_matrix.

In addition, the whole example that uses the plotters module has
also been removed due to deprecation.

* Update plot_elbow_curve example notebook (#62)

This commit updates the plot_elbow curve example where it removes
the need to create a clusterer factory and instead uses the
scikitplot.cluster.plot_elbow_curve method.

* Update plot_feature_importance notebook (#62)

This commit updates the use of the plot_feature_importance
method to be compliant with v.0.3.0. The additional example
that uses plotters has been removed for it is deprecated already.

* Update plot_feature_importance

* Update plot_ks_statistic notebook example (#62)

This commit updates the example notebook by using the
scikitplot.metrics.plot_ks_statistic API. This also removes
the plotter example due to depreciation.

* Update plot_learning_curve notebook example (#62)

* Update plot_pca_2d_projection example notebook (#62)

* Update plot_pca_2d_projection example notebook (#62)

* Update plot_precision_recall_curve example (#62)

* Define colormaps explicitly in examples

This commit defines the colormap explicitly in the plot_precision_recall
example. It turns out that for other matplotlib versions, a warning on
nipy_spectral/nipy_spectral_r is being raised. Defining it avoids this
ugly warning.

* Update plot_roc_curve examples (#62)

This commit updates plot_roc_curve to be compliant with
version 0.3.0. Moreover, the colormap here is defined
explicitly in the function call in order to avoid the ugly
warning raised by nipy_spectral.

* Update description on example

* Update plot_silhouette notebook example (#62)